+++ CHANGES.md
@@ -1,6 +1,7 @@
 # v1.0.0
 - Refresh orders
 - Order orders by timestamp
+- Create Driver tag to receive notifications
 
 # v0.0.2
 - Add `ready` orders tab